+++ tools.py
@@ -45,6 +45,64 @@ def clear_dir(dir_path):
 FILE TOOLS METHOD
 """
 
+"""
+BACKUP CONVINENCE METHODS TOOLS
+"""
+def do_backup():
+    #check if backup folder exists
+    backup_path = config["backup-path"]
+    secured_cred_path = config["cred-output-dir"]
+    if not os.path.exists(backup_path):
+        os.makedirs(backup_path)
+    #copy all and decrypt files in secured credentials to backup folder
+    message_info("Backing up files")
+    encrypted_file_list = list_files_in_dir(secured_cred_path)
+    if encrypted_file_list is not None:
+        import core
+        encrypted_file_list.remove(".gitignore")
+        message_info("Creating Backup")
+        
+        #Decrypt the files and move it into backup
+        for file in encrypted_file_list:
+            try:
+                data = core.read_encrypted_file(file)
+                #strip "encrypted-" from filename
+                filename = file.replace("encrypted-", "")
+                makeCopyOfFile(oldFileName = filename, newContent= data, status="backup", path=backup_path)
+            except Exception as e:
+                message_warn(e)
+                print()
+        
+        message_info("Succesfully doing backup")
+     
+        # clear_dir(backup_path)
+def import_backup():
+    message_info("Importing backup files")
+    backup_path = config["backup-path"]
+    
+    backupFileList = list_files_in_dir(backup_path)
+    cred_path = config["cred-input-dir"]
+    if backupFileList is not None:
+        message_info(backupFileList)
+        for file in backupFileList:
+            try:
+                file_path = backup_path + "\\" + file
+                data = readFileContent(file_path)
+
+                #remove the state from original filename
+                filename = file.replace("backup-", "")
+                # move file into cred-input-dir
+                with open(cred_path + "\\" + filename, "w") as f:
+                    f.write(data)                
+            except Exception as e:
+                message_warn(e)     
+        message_info("Succesfully imported backup files")
+        #remove backup dir
+        message_info("Removing backup directory")
+        os.popen("rd /s /q " + backup_path)  
+"""
+BACKUP CONVINENCE METHODS TOOLS
+"""
 
 """
 DEPENDENCIES TOOLS
@@ -75,6 +133,7 @@ def satisfy_dependencies():
     message_info("succesfully cloned all required dependencies repos")
 
 def update():
+    
     loop = asyncio.get_event_loop()
     try:
         loop.run_until_complete(pull(load_config("config.json")["cipher_path"]))
@@ -84,6 +143,8 @@ def update():
         print()#space
     finally:
         message_info("Please re-run the program to apply changes")
+        
+    
 """
 DEPENDENCIES TOOLS
 """
